feat: Updated utils/check_self_hosted_runner.py

+++ utils/check_self_hosted_runner.py
@@ -21,6 +21,11 @@ def get_runner_status(target_runners, token):
                 offline_runners.append(runner)
 
     # save the result so we can report them on Slack
+    import os
+    if not os.path.exists("offline_runners.txt"):
+        print("No offline runners.")
+        return
+
     with open("offline_runners.txt", "w") as fp:
         fp.write(json.dumps(offline_runners))
